+++ CHANGELOG.md
@@ -27,6 +27,7 @@ arbitrarily applied mapper pass.
 - `CircuitAnalyzer` analyzer pass for computing structural circuit metrics (size, interaction graph, gate dependency graph, density)
 - `Can2CZDecomposer` to decompose arbitrary two-qubit gates.
 - The following 2-qubit gates: `CV`, `CY`, `DCNOT`, `ECR`, `ISWAP`, `InvSqrtSWAP`, `M`, `MS`, `SqrtISWAP`, and `SqrtSWAP`
+- Add `add_instruction` method to the `CircuitBuilder`
 
 ## [ 0.9.0 ] - [ 2025-12-19 ]